% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_jumble_tree.R
\name{best_jumble_tree}
\alias{best_jumble_tree}
\title{best_jumble_tree}
\usage{
best_jumble_tree(tree_dir_path, multiphylo)
}
\arguments{
\item{tree_dir_path}{Path to a directory containing multiple jumble trees}

\item{multiphylo}{A multiphylo object containing multiple jumble trees}
}
\value{
NULL Prints to consolethe index of the tree with the maximum sum, mean and median of jumble scores. To identify which tree file the index corresponds to use 'list.files(tree_dir_path)'.
}
\description{
Identify the jumeble tree with the highest sum, mean and median of jumble scores.
}
