% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_tree.R
\name{color_tree}
\alias{color_tree}
\title{color_tree}
\usage{
color_tree(tree_path, print)
}
\arguments{
\item{tree_path}{Path to tree}

\item{print}{Print out the colored trees as pdf}
}
\value{
A list containing two trees: 'tip_tree' = the tree with tip labels colored by cell type and 'bar_tree': the tree with cell types indicated with a colored bar
}
\description{
color_tree
}
