% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_matrices.R
\name{combine_matrices}
\alias{combine_matrices}
\title{combine_matrices}
\usage{
combine_matrices(matrix_list, ortholog_map_path, combined_mat_name)
}
\arguments{
\item{matrix_list}{A list of the full paths to the matrices you want to combine. Cell ids must be in cellphylo format.}

\item{ortholog_map_path}{Path to wrangled Ensembl orthologs file. The custom script for wrangling the raw Ensembl orthologs file can be found in wrangle_Ensembl_orthologs.Rmd.}

\item{combined_mat_name}{Name for the combined matrix this function produces. Cell ids are in cellphylo format.}
}
\value{
full_join_sparse The cross-species combined matrix in sparse matrix format
}
\description{
Takes in multiple species matrices and combines them into a single multi-species matrix combined by gene id. This combined matrix is NOT cross-species normalized yet.
}
