% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_10x.R
\name{convert_to_10x}
\alias{convert_to_10x}
\title{convert_to_10x}
\usage{
convert_to_10x(path_to_csv, species_name)
}
\arguments{
\item{path_to_csv}{Path to csv file}

\item{species_name}{Name of species matrix belongs to.}
}
\value{
sparse The matrix in sparse matrix format and a directory with a matrix in 10x feature-barcode format (version 3)
}
\description{
Converts a UMI count matrix in CSV format to 10x-style feature-barcode matrices.
}
