% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_replicates.R
\name{count_replicates}
\alias{count_replicates}
\title{count_replicates}
\usage{
count_replicates(path_to_dir)
}
\arguments{
\item{path_to_dir}{Path to 10x matrix directory. Cell ids must be in cellphylo format.}
}
\description{
count_replicates
}
