% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species_integration.R
\name{cross_species_integration}
\alias{cross_species_integration}
\title{cross_species_integration}
\usage{
cross_species_integration(matrix_path)
}
\arguments{
\item{matrix_path}{Path to combined cross-species, unintegrated matrix produced by cellphylo::combine_matrices. Cell ids must be in cellphylo format.}
}
\value{
combined The Seurat object containing the cross-species integrated matrix.
}
\description{
cross_species_integration
}
