% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene_loadings.R
\name{extract_gene_loadings}
\alias{extract_gene_loadings}
\title{extract_gene_loadings}
\usage{
extract_gene_loadings(ann_table, matrix_path)
}
\arguments{
\item{ann_table}{Annotation table connecting gene ids to human gene symbols.}

\item{matrix_path}{Path to matrix in 10x format to do PCA on and extract gene loadings.}
}
\value{
gene_loadings_output A list containing 1) 'by_most_Positive': a data frame with gene symbols have been ordered by most positive loading for each PC 2) "by_most_Negative": A data frame with gene symbol have been ordered by mmost negative loading for each PC, "PCA_output": the object produced by performing PCA on the input matrix with prcomp.
}
\description{
extract_gene_loadings
}
