% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_errored_runs.R
\name{id_errored_runs}
\alias{id_errored_runs}
\title{id_errored_runs}
\usage{
id_errored_runs(outtrees_dir, matrix_name_pattern)
}
\arguments{
\item{outtrees_dir}{Path to directory with contml \code{outtrees}. This directory contains all \code{outtree} files created by contml, including for runs that errored.}

\item{matrix_name_pattern}{Prefix for matrix name. This function assumes that your matrix names end in \code{index_num_mtx} eg. for \code{scjackknife_540_TBE_1_mtx}, the matrix_name_pattern is \code{scjackknife_540_TBE}}
}
\value{
prints out a list of scjackknife matrix names that produced empty tree files when run through contml.
}
\description{
When \code{contml} runs error, it produces an empty \code{outtree} and \code{outfile}. We remove these, but now the index between the \code{contml} output files and scjack matrices do not match, leading to downstream analysis issues.
This function makes a list of matrices that produced an error when \code{contml} was run. We will then use this list to remove these matrices and proceed with downstream analyses.
This function is part of the workflow of the scjackknife analysis.
}
