% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_scjack_trees.R
\name{label_scjack_trees}
\alias{label_scjack_trees}
\title{label_scjack_trees}
\usage{
label_scjack_trees(
  n_trees,
  matrix_dir_path,
  tree_dir_path,
  print = TRUE,
  file_name_prefix
)
}
\arguments{
\item{n_trees}{The number of trees to be relabelled.}

\item{matrix_dir_path}{The path to the scjackknife matrices that were used to create the scjackknife infiles. This directory must have matrices that produced empty outtree files removed using cellphylo::id_errored_runs.}

\item{tree_dir_path}{The path to the scjakknife outtrees produced by contml. This must have empty outtree files removed.}

\item{print}{Boolean. Print out trees with 1) full labels and 2) common labels}

\item{file_name_prefix}{The prefix for the outtree files to produce.}
}
\value{
output A list containing scjackknife trees with full tip name labels ("labelled_tree") and scjackknife trees with comomn labels ("common_labels_tree)
}
\description{
Comparison and calculation of TBE across scjackknife trees require that the tips share identical names
This script labels scjackknife outtrees by 1) their full tip names and 2) a common label - a tip name that contains only species and cell type cluster group id that can be matched acros scjackknife trees
Previously named label_scjack_trees_54.
}
