% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tree.R
\name{label_tree}
\alias{label_tree}
\title{label_tree}
\usage{
label_tree(matrix_path, tree_path, file_name, print)
}
\arguments{
\item{matrix_path}{Path to matrix from which the contml infile was created. The taxon numbers of the infile correspond to the matrix cell ids in sequence.}

\item{tree_path}{Path to the raw phylogeny produced by contml ("outtree" file)}

\item{file_name}{Name of tree file of labelled tree}

\item{print}{Print out the tree file of the labelled tree}
}
\value{
A phylo object of the labelled tree
}
\description{
Replace the taxon number labels in the contml outtree with the cell ids in the input matrix used to make the contml infile.
}
