% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pc_sweep_infiles.R
\name{make_pc_sweep_infiles}
\alias{make_pc_sweep_infiles}
\title{make_pc_sweep_infiles}
\usage{
make_pc_sweep_infiles(matrix_path)
}
\arguments{
\item{matrix_path}{Path to matrix. For Mah and Dunn (2023) use \code{aqhumor_cross_species_integrated_mtx}}
}
\description{
make_pc_sweep_infiles
}
