% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scjack_sample.R
\name{make_scjack_sample}
\alias{make_scjack_sample}
\title{make_scjack_sample}
\usage{
make_scjack_sample(
  n_samples,
  subsample_matrix_path,
  full_matrix_path,
  file_name
)
}
\arguments{
\item{n_samples}{Number of scjackknife samples to create.}

\item{subsample_matrix_path}{Path to the matrix to sample from.}

\item{full_matrix_path}{Path to full unsubsetted matrix. PCA should already have been performed and PCs subsetted.}
}
\value{
pca_matrix List of resampled single jackknife matrices.
}
\description{
Create resampled matrices that contain 1 replicate cell per cell type group from a larger matrix that contains multiple replicates. This is used for running the scjackknife analysis.
Previously named make_scjack_sample54
}
