% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scjack_sample_54.R
\name{make_scjack_sample_54}
\alias{make_scjack_sample_54}
\title{make_scjack_sample_54
Create resampled matrices that contain 1 replicate cell per cell type group from a larger matrix that contains multiple replicates. This is used for running the scjackknife analysis.}
\usage{
make_scjack_sample_54(n_samples, matrix_path)
}
\arguments{
\item{n_samples}{Number of scjackknife samples to create.}

\item{matrix_path}{Path to the matrix to sample from.}
}
\value{
pca_matrix List of resampled single jackknife matrices.
}
\description{
make_scjack_sample_54
Create resampled matrices that contain 1 replicate cell per cell type group from a larger matrix that contains multiple replicates. This is used for running the scjackknife analysis.
}
