% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jumble.R
\name{score_jumble}
\alias{score_jumble}
\title{score_jumble}
\usage{
score_jumble(multiphylo, focal_tree_path, focal_tree_obj, print)
}
\arguments{
\item{multiphylo}{A multiphylo object containing the jumble trees}

\item{focal_tree_path}{Path to the tree to be scored.}

\item{focal_tree_obj}{An object of class phylo containing the tree to be scored.}

\item{print}{Boolean. Print out the scored tree and 50\% consensus tree}
}
\value{
output List object containing the scored tree, the 50\% consensus tree, and the sum, mean and median of the scores of the focal tree
}
\description{
Calculate the jumble score from a collection of jumble trees and plot onto a focal tree. The jumble score is the clade frequency of each split across jumble trees.
}
