% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_scjack.R
\name{score_scjack}
\alias{score_scjack}
\title{score_scjack}
\usage{
score_scjack(scjack_multiphylo, focal_tree_path, focal_tree_obj, print)
}
\arguments{
\item{scjack_multiphylo}{A multiphylo object containing the jumble trees}

\item{focal_tree_path}{Path to the tree to be scored.}

\item{focal_tree_obj}{An object of class phylo containing the tree to be scored.}

\item{print}{Boolean. Print out the scored tree and 50\% consensus tree}
}
\value{
output List object containing the scored tree, the 50\% consensus tree, and the sum, mean and median of the scores of the focal tree
}
\description{
score_scjack
}
