% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_cells.R
\name{subsample_cells}
\alias{subsample_cells}
\title{subsample_cells}
\usage{
subsample_cells(path_to_matrix, n_reps)
}
\arguments{
\item{path_to_matrix}{Path to matrix you want to subsample from. Cell ids must be in cellphylo format.}

\item{n_reps}{}
}
\value{
mat.sub A matrix created by random subsampling of the input matrix. All cell ids must be in cellphylo format.
}
\description{
subsample_cells
}
