% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_cells.R
\name{subset_cells}
\alias{subset_cells}
\title{subset_cells}
\usage{
subset_cells(path_to_matrix, n_reps, print = TRUE)
}
\arguments{
\item{path_to_matrix}{Path to matrix you want to subsample from. Matrix must be in cellphylo format.}

\item{n_reps}{Number of replicates per cell type group to be in the subsampled matrix.}
}
\value{
mat.sub A sparse matrix created by random subsampling of the input matrix. If print = TRUE, prints out the cell ids of the sampled replicates (subset_cells_list) and the subsampeld matrix in 10x feature-barcode format.
}
\description{
This function randomly samples n replicates per cell type group from a larger counts matrix. The input matrix must be in cellphylo format.
}
