% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_combined_matrix.R
\name{subset_combined_matrix}
\alias{subset_combined_matrix}
\title{subsample_combined_matrix}
\usage{
subset_combined_matrix(matrix_path, sample_size, print)
}
\arguments{
\item{matrix_path}{Path to a combined matrix of cells from multiple species.}

\item{sample_size}{The number of replicates per cell type group to be sampled}

\item{print}{Print out a file listing the cell ids of selected cells and the subsampled matrix in 10x format.}
}
\value{
mat.sparse The subsampled matrix in sparse matrix format.
}
\description{
Randomly subsample a combined matrix containing cells from multiple species. Cell ids must be in cellphylo format.
}
