% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_species_integr.R
\name{within_species_integr}
\alias{within_species_integr}
\title{within_species_integr}
\usage{
within_species_integr(path_to_matrix)
}
\arguments{
\item{path_to_matrix}{Path to a species' matrix. Cell ids must be in cellphylo format.}
}
\value{
NULL RDS file of integrated Seurat object and 3 matrices in 10x format featuring values from the integrated Seurat object SCT assay "scale.data" slot (sct_pearson_residuals), SCT assay "counts" slot (sct_corrected_UMI_counts), and integrated assay "scale.data" slot (sct_integrated_pearson_residuals)
}
\description{
within_species_integr
}
