% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_species_integration.R
\name{within_species_integration}
\alias{within_species_integration}
\title{within_species_integration}
\usage{
within_species_integration(
  path_to_matrix,
  n_features = 3000,
  k_weight = 100,
  print = TRUE
)
}
\arguments{
\item{path_to_matrix}{Path to a species' matrix. Cell ids must be in cellphylo format.}

\item{n_features}{The number of features to use for integration using Seurat::SelectIntegrationFeatures. Default is 3000, used in Mah & Dunn 2023.}

\item{k_weight}{The k.weight to use with Seurat::IntegrateData. This must be less than or equal to the number of cells in the smallest batch. Default is 100, used in Mah & Dunn 2023.}

\item{print}{Boolean. Print out the Seurat object as an RDS file and extract and print out the pearson residuals, corrected UMI counts and integrated pearson residuals from the Seurat object. These files will be deposited in \code{matrix/within-species_analysis}.}
}
\value{
combined.sct A Seurat object featuring the normalized and batch-integrated matrices.
If print=TRUE, it will also print out:
An RDS file of the Seurat object and 3 matrices in 10x format featuring values from the normalized and batch-integrated Seurat object: 1. sct_pearson_residuals (SCT assay "scale.data" slot), 2; sct_corrected_UMI_counts (SCT assay "counts" slot), 3.sct_integrated_pearson_residuals (integrated assay "scale.data" slot)
}
\description{
This function normalizes counts using a negative binomial model (SCTransform) and performs within-species batch integration using the Seurat workflow. Matrices must be in cellphyo format.
}
