% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_matrices.R
\name{combine_matrices}
\alias{combine_matrices}
\title{combine_matrices}
\usage{
combine_matrices(matrix_list, ortholog_map_path, print = TRUE)
}
\arguments{
\item{matrix_list}{A list of the full paths to the matrices you want to combine. Cell ids must be in cellphylo format.}

\item{ortholog_map_path}{Path to wrangled Ensembl orthologs file.}

\item{print}{Boolean. Print out the combined matrix.}
}
\value{
full_join_sparse The combined matrix in sparse matrix format. If print=TRUE, it will print out the combined matrix and deposit it in \code{matrix/cross-species_analysis/combined_by_orthology}
}
\description{
Takes in multiple species' matrices and combines them into a single multi-species matrix combined by gene orthology. This combined matrix is NOT cross-species normalized yet.
}
