% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_multi_trees.R
\name{combine_multi_trees}
\alias{combine_multi_trees}
\title{combine_multi_trees}
\usage{
combine_multi_trees(tree_dir_path)
}
\arguments{
\item{tree_dir_path}{Path to directory containing multiple trees to combine}
}
\value{
multi_tree A multiphylo object of all trees in the tree_dir_path directory
}
\description{
combine_multi_trees
}
