% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_10x.R
\name{convert_to_10x}
\alias{convert_to_10x}
\title{convert_to_10x}
\usage{
convert_to_10x(path_to_csv, dir_name)
}
\arguments{
\item{path_to_csv}{Path to csv file}

\item{dir_name}{Name for 10x matrix directory}
}
\value{
NULL a directory with a matrix in 10x format (version 3)
}
\description{
convert_to_10x
}
