% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species_integration.R
\name{cross_species_integration}
\alias{cross_species_integration}
\title{cross_species_integration}
\usage{
cross_species_integration(matrix_path, k_weight = 100, print = TRUE)
}
\arguments{
\item{matrix_path}{Path to combined matrix. This combined matrix features cells from multiple species but is not yet cross-species integrated. This matrix was created using \code{cellphylo::combine_matrices}. Cell ids must be in cellphylo format.}

\item{k_weight}{The k.weight to use with Seurat::IntegrateData. This must be less than or equal to the number of cells in the smallest batch. Default is 100.}

\item{print}{Boolean. Print out the matrix and Seurat object of cross-species integrated data.}
}
\value{
combined The Seurat object containing the cross-species integrated matrix. Print=TRUE prints out the cross-species integrated matrix and an RDS file of the Seurat object that results from the Seurat integration workflow.
}
\description{
Integrate counts across species using the Seurat workflow. Parameters used in Mah & Dunn (2023) in the Seurat workflow are hard-coded into this function.
}
