% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_errored_runs.R
\name{id_errored_runs}
\alias{id_errored_runs}
\title{id_errored_runs}
\usage{
id_errored_runs(outtrees_dir)
}
\arguments{
\item{outtrees_dir}{Path to directly with contml outtrees.}
}
\value{
NULL prints out a list of scjackknife matrix names that produced empty tree files when run through contml.
}
\description{
Create a list of matrices produced by contml runs that errored out.
Errored contml runs lead to empty outtree files, which must be removed for downstream analyses. However, the corresponding scjackknife matrices must also be removed to preserve index number in downstream analyses.
This function is part of the workflow of the scjackknife analysis.
}
