% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_infile.R
\name{make_infile}
\alias{make_infile}
\title{make_infile}
\usage{
make_infile(matrix, matrix_path, print)
}
\arguments{
\item{matrix}{Matrix to create contml infile from}

\item{matrix_path}{Path to matrix to create contml infile from}

\item{print}{Print the infile}
}
\value{
An object representing a matrix formatted in contml infile format.
}
\description{
Converts a matrix to contml (PHYLIP) infile format.
}
