% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eigs.R
\name{plot_eigs}
\alias{plot_eigs}
\title{plot_eigs}
\usage{
plot_eigs(matrix_path)
}
\arguments{
\item{matrix_path}{Path to matrix PCA is to be performed on.}
}
\description{
Perform PCA and produce eigenvalue plots for a matrix.
}
