% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_sweep.R
\name{plot_pc_sweep}
\alias{plot_pc_sweep}
\title{plot_pc_sweep}
\usage{
plot_pc_sweep(rank = c(1:98), var_df, print = TRUE)
}
\arguments{
\item{rank}{The range of the ranks you want to plot. This corresponds to row number in var.df calcualted by calc_pc_sweep_var.}

\item{var_df}{The dataframe of variable produced by cellphylo::calc_pc_sweep_var}
}
\value{
List containing plots for tree length, average edge length and star-ness score
}
\description{
Plots the variables calculated by calc_pc_sweep_var. Produces Figs 2A, 2B, 2C.
}
