% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_sweep.R
\name{plot_pc_sweep}
\alias{plot_pc_sweep}
\title{plot_pc_sweep}
\usage{
plot_pc_sweep(rank, var_df)
}
\arguments{
\item{rank}{The rank index of the trees to plot. This must match the trees submitted as input to calc_pc_sweep_var. Eg. If you want to plot trees inferred from PC sweep matrices 3-100, rank is c(3:100).}

\item{var_df}{The dataframe of variable produced by cellphylo::calc_pc_sweep_var}
}
\value{
List containing plots for tree length, average edge length and star-ness score
}
\description{
plot_pc_sweep
}
