% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_matrix.R
\name{wrangle_matrix}
\alias{wrangle_matrix}
\title{wrangle_matrix}
\usage{
wrangle_matrix(path_to_mat, metadata_file, species)
}
\arguments{
\item{path_to_mat}{Path to matrix to be wrangled into cellphylo format. Matrix must be in 10x format.}

\item{metadata_file}{Path to wrangled metadata file from van Zyl et al. (2020)}

\item{species}{Name of species the matrix belongs to}
}
\value{
mat A wrangled matrix with cell ids in cellphylo format
}
\description{
Converts cell identifiers from van Zyl et al. (2020)'s format to cellphylo format.
}
