% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_metadata.R
\name{wrangle_metadata}
\alias{wrangle_metadata}
\title{wrangle_metadata}
\usage{
wrangle_metadata(path_to_meta_file)
}
\arguments{
\item{path_to_meta_file}{A path to the metadata file provided by van Zyl et al. (2020).}
}
\value{
meta_data A new file with metafile data parsed in preparation of creating cellphylo cell ids
}
\description{
wrangle_metadata
}
